pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cookiejack"
version = "0.2"
description = "Sniff cookies off the wire and send them via a websocket."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Ed Cradock" },
]
dependencies = [
	"attrs ~= 25.1.0",
	"autobahn ~= 24.4.2",
	"click ~= 8.1.8",
	"scapy ~= 2.6.1",
	"Twisted ~= 24.11.0",
	"zope.interface ~= 7.2",
]

[project.scripts]
cookiejack = "cookiejack.cli:main"
